// FE/src/api/models/createServiceGroupDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * SIGREF.API
 * OpenAPI spec version: 1.0
 */
import type { ListStatus } from "./listStatus";
import type { CodeableConceptDto } from "./codeableConceptDto";

export interface CreateServiceGroupDto {
  /** @minLength 1 */
  title: string;
  status?: ListStatus;
  code?: CodeableConceptDto;
  /** @nullable */
  healthcareServiceIds?: string[] | null;
  /** @nullable */
  locationIds?: string[] | null;
}